feat(add_new_zodiac_type): add tropical zodiac type
