+++ lib/signs/sidereal.json
@@ -1,7 +1,7 @@
 [
   {
-    "min": "July 17",
-    "max": "August 6",
+    "min": "January 20",
+    "max": "Febuary 13",
     "sign": "Cancer",
     "facts": ["Intuitive", "Emotional", "Intelligent", "Passionate"],
     "element": "Water",
@@ -74,7 +74,7 @@
     "min": "February 14",
     "max": "March 9",
     "sign": "Aquarius",
-    "facts": [],
+    "facts": ["Independent", "Open-Minded", "Inventive"],
     "element": "Air",
     "luckyGem": "Amethyst"
   },
