+++ lib/tell-sign.js
@@ -1,32 +1,10 @@
-const signs = require('./signs.json');
 const Util = require('../utils');
+const { getPossibleZodiacs, getZodiacDates, getZodiacSign } = require('./helpers');
 
-module.exports = (birthdayDate) => {
+module.exports = (birthdayDate, type = 'sidereal') => {
+  const signs = require(type === 'sidereal' ? './signs/sidereal.json' : './signs/tropical.json');
   const [birthMonthStr, birthDay] = Util.getDateDetails(birthdayDate);
-  const filteredSigns = getPossibleZodiacs(signs, birthMonthStr);
-  const signDetails = getZodiacSign(filteredSigns, birthMonthStr, birthDay);
+  const filteredSigns = getPossibleZodiacs(signs, birthMonthStr, getZodiacDates);
+  const signDetails = getZodiacSign(filteredSigns, { birthMonthStr, birthDay, getZodiacDates });
   return signDetails;
-}
-
-
-function getZodiacSign(signs, birthMonthStr, birthDay) {
-  const signDetails = signs.find(sign => {
-    const [monthMin, monthMax, dayMin, dayMax] = getZodiacDates(sign);
-    return (birthDay >= dayMin && birthMonthStr === monthMin) || (birthDay <= dayMax && birthMonthStr === monthMax);
-  });
-
-  return signDetails;
-}
-
-function getZodiacDates(sign) {
-  const [monthMin, dayMin] = sign.min.split(' ');
-  const [monthMax, dayMax] = sign.max.split(' ');
-  return [monthMin, monthMax, dayMin, dayMax];
-}
-
-function getPossibleZodiacs(signs, birthMonthStr, birthDay) {
-  return signs.filter((sign) => {
-    const [monthMin, monthMax] = getZodiacDates(sign);
-    return monthMin === birthMonthStr || monthMax === birthMonthStr;
-  });
 }
