+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sign-teller",
-  "version": "1.0.2",
+  "version": "2.0.0",
   "main": "lib/index.js",
   "repository": {
     "type": "git",
